+++ icevision/models/efficientdet/__init__.py
@@ -7,7 +7,7 @@ from icevision.models.efficientdet.show_results import *
 # Soft dependencies
 from icevision.soft_dependencies import SoftDependencies
 
-if SoftDependencies.fastai2:
+if SoftDependencies.fastai:
     import icevision.models.efficientdet.fastai
 
 if SoftDependencies.pytorch_lightning: